GoodCity May 2017 release

+++ app/jobs/cloudinary_image_cleanup_job.rb
@@ -2,6 +2,8 @@ class CloudinaryImageCleanupJob < ActiveJob::Base
   queue_as :low
 
   def perform(cloudinary_image_id)
-    Cloudinary::Api.delete_resources([cloudinary_image_id])
+    unless %w(development test).include?(Rails.env)
+      Cloudinary::Api.delete_resources([cloudinary_image_id])
+    end
   end
 end
